Refactor matterclient
* Split up in different files
* Decrease complexity
